% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{bold}
\alias{bold}
\title{FMRI BOLD data}
\description{
Dataset contains fMRI BOLD data from from five subjects in each of five
resting-state conditions. Object is a data frame with the following fields:
\itemize{
  \item{"Subject"} Subject ID
  \item{"Condition"} Condition label
  \item{"Data"} A 150 x 5 matrix containing the standardized BOLD signal
  from 5 cortical regions.
}
}
\author{
Corson N. Areshenkoff \email{areshenk@protonmail.com}
}
\keyword{data}
