% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-spd.R
\name{is.spd}
\alias{is.spd}
\title{Check if a matrix is symmetric and positive definite}
\usage{
is.spd(..., tol = 1e-08)
}
\arguments{
\item{...}{One or more numeric matrices}

\item{tol}{Numerical tolerance for checking symmetry and positive definiteness}
}
\value{
A Boolean value indicating whether the matrix is symmetric and positive
definite
}
\description{
Check if a matrix is symmetric and positive definite
}
\details{
Function checks whether the matrix \code{x} is symmetric and positive
definite. Symmetry is evaluated up to an entrywise tolerance of \code{tol}, so
that differences smaller than \code{tol} are ignored. Positive-definiteness is
checked by computing the eigenvalues of \code{x} using \code{eigen}, setting
eigenvalues smaller than \code{tol} in absolute value to zero, and then checking
whether any are less than or equal to zero.
}
