% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-correlation.R
\name{spd.correlation}
\alias{spd.correlation}
\title{Compute (partial) correlations}
\usage{
spd.correlation(x, method = "correlation")
}
\arguments{
\item{x}{An SPD matrix to be whitened}

\item{method}{A string specifying either "correlation" (default) orf
"partial" correlation.}
}
\value{
A symmetric, positive-definite matrix.
}
\description{
Transforms an SPD matrix into a matrix of partial correlations
}
