% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-dist.R
\name{spd.dist}
\alias{spd.dist}
\title{Distance between two symmetric, positive-definite matrices}
\usage{
spd.dist(x, y, method = "euclidean", ...)
}
\arguments{
\item{x, y}{Symmetric, positive-definite matrices}

\item{method}{The distance measure. See details.}
}
\description{
Function implements several distance measures between symmetric,
positive-definite matrices.
}
\details{
Allowable distance measures are
\itemize{
 \item{"euclidean": }{The Frobenius norm of the difference \code{x-y}.}
 \item{"logeuclidean": }{The Frobenius norm of the difference \code{log(x)-log(y)} in the tangent space.. }
 \item{"cholesky": }{The Frobenius norm of the difference between the cholesky factors
 of \code{x} and \code{y}. Not affinely invariant.}
 \item{"riemannian": }{The Riemmanian distance proposed by Barachant, et al. (2013)}
 \item{"stein": }{The square root of Jensen-Bregman log determinant divergence}
}
}
