% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-expmap.R
\name{spd.expmap}
\alias{spd.expmap}
\title{Projection from the tangent space}
\usage{
spd.expmap(x, p = NULL)
}
\arguments{
\item{x}{A symmetric matrix}

\item{p}{The point to whose tangent space x belongs}
}
\value{
A symmetric, positive-definite matrix.
}
\description{
Function projects a tangent vector x (a symmetric matrix) at a point p
onto the space of SPD matrices.
}
