% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.fa.R
\name{spd.fa}
\alias{spd.fa}
\title{Fractional anisotropy of a symmetrix, positive-definite matrix}
\usage{
spd.fa(x)
}
\arguments{
\item{x}{A symmetric, positive-definite matrix}
}
\description{
Fractional anisotropy of a symmetrix, positive-definite matrix
}
