% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-interp.R
\name{spd.interpolate}
\alias{spd.interpolate}
\title{Interpolation between two symmetric, positive-definite matrices}
\usage{
spd.interpolate(x, y, t, method = "euclidean", ...)
}
\arguments{
\item{x, y}{Symmetric, positive-definite matrices}

\item{t}{Interpolation parameter in \code{[0,infinity)}}

\item{method}{Type of interpolation (see details)}
}
\description{
Function smoothly interpolates between two SPD matrices \code{x} and \code{y}.
The interpolation is parametrized by \code{t}, where \code{t=0} returns \code{x},
\code{t=1} returns \code{y}, and \code{t=.5} returns the mean of \code{x} and \code{y}
}
\details{
Allowable distance measures are
\itemize{
 \item{"euclidean": }{Euclidean interpolation \code{(1-t)x + ty}}
 \item{"logeuclidean": }{Euclidean interpolation in the tangent space.}
 \item{"riemannian": }{Interpolation along the geodesic path from \code{x} to \code{y}}
}
}
