% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-logmap.R
\name{spd.logmap}
\alias{spd.logmap}
\title{Projection onto the tangent space}
\usage{
spd.logmap(x, p = NULL)
}
\arguments{
\item{x}{A symmetric positive definite matrix}

\item{p}{The point on whose tangent space to project x}
}
\value{
A symmetric matrix.
}
\description{
Function projects an SPD matrix x onto the tangent space at a point p
}
