% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-pca-predict.R
\name{spd.pca.predict}
\alias{spd.pca.predict}
\title{Prediction for SPD kernel PCA}
\usage{
spd.pca.predict(fit, x)
}
\arguments{
\item{fit}{An object of class kpca return by spd.pca}

\item{x}{A list of SPD matrices for which to derive component scores.}
}
\value{
A matrix of principal component scores
}
\description{
Function accepts a fitted kpca object returned by \code{spd.pca} and a list of
SPD matrices and returns a matrix of estimated principal component scores.
}
