% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-svm.R
\name{spd.svm}
\alias{spd.svm}
\title{Kernel svm for SPD matrices}
\usage{
spd.svm(x, y, p, test.idx, ...)
}
\arguments{
\item{x}{A list of symmetric, positive-definite matrices}

\item{y}{A vector of labels}

\item{p}{The basepoint, on whose tangent space to project the elements of x}

\item{test.idx}{A vector of indices indicating the element of x to be witheld
for testing}

\item{...}{Further arguments for kernlab::ksvm.}
}
\value{
A data.frame with test set labels and predictions.
}
\description{
Functions trains a kernel SVM to a list of SPD matrices, and generates
predictions on a specified test set. The kernel matrix is constructed
by projecting the observations onto the tangent space at a basepoint p
and taking the inner-product given by the Riemannian metric. This can
be interpreted roughly as linearizing the space of SPD matrices around p
and fitting a linear svm. This function is more or less a wrapper around the
kernlab function \code{ksvm}.
}
