% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-translate.R
\name{spd.translate}
\alias{spd.translate}
\title{Translation of a covariance matrix}
\usage{
spd.translate(x, from, to = NULL, ...)
}
\arguments{
\item{x}{A symmetric, positive-definite matrix}

\item{from}{The SPD matrix to whose tangent space x belong}

\item{to}{The SPD matrix to whose tangent space to move x}

\item{...}{Additional arguments to spd.transport}
}
\value{
A symmetric, positive-definite matrix
}
\description{
Translates a covariance matrix from one basepoint to another, Formally,
projects a covariance matrix onto the tangent space at one basepoint,
transports to the tangent space at the second basepoint, and projects back onto
the space of SPD marices.
}
