% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-transport.R
\name{spd.transport}
\alias{spd.transport}
\title{Transport a tangent vector}
\usage{
spd.transport(x, from, to = NULL, method = "gl", nsteps = NA)
}
\arguments{
\item{x}{A tangent vector (a symmetric matrix)}

\item{from}{The SPD matrix to whose tangent space x belongs}

\item{to}{The SPD matrix to whose tangent space to move x}

\item{method}{See details}

\item{nsteps}{For parallel transport, the number of steps along the geodesic
connecting from and to}
}
\value{
A symmetric matrix -- a tangent vector at to.
}
\description{
Transports a tangent vector from one tangent space to another.
}
\details{
Allowable method are
\itemize{
 \item{"pt": }{Parallel transport using the Schild's ladder algorithm}
 \item{"gl": }{Transport from A to B using the GL(n) action \eqn{A -> GAG'}, where
 \eqn{G = B^{1/2}A^{-1/2}}}
}
}
