% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-whiten.R
\name{spd.whiten}
\alias{spd.whiten}
\title{Covariance whitening transform}
\usage{
spd.whiten(x, p, unwhiten = F)
}
\arguments{
\item{x}{An SPD matrix to be whitened}

\item{p}{The SPD baseline whitening matrix}

\item{unwhiten}{Logical. If TRUE, reverses a previously applied whitening transform.}
}
\value{
A symmetric, positive-definite matrix.
}
\description{
Whitens an SPD matrix using the procedure advocated by Ng, et al. (2014)
}
