% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.center.R
\name{spd.center}
\alias{spd.center}
\title{Center a set of spd matrices}
\usage{
spd.center(x, y = NULL)
}
\arguments{
\item{x}{Either a list of symmetric, positive-definite matrices, an individual amtrix}

\item{y}{Value(s) for centering. Either a list of spd matrices of the same
length as x, a single spd matrix, or NULL. If NULL, uses the mean of \code{x}}
}
\value{
An object of the same type as \code{x}, containing the centered matrices..
}
\description{
Functions centers a set of symmetric, positive-definite matrices so that they
have mean equal to the identity matrix.
}
\details{
Functions transports the matrices in \code{x} so that \code{y} lies at
the identity. If \code{x} is a list and \code{y = NULL} (default), function sets
\code{y} equal to the mean of \code{x}. Otherwise, \code{y} can be set to any
spd matrix. When \code{x} is a matrix, \code{y} must be provided. If \code{y}
is a list of the same length as {x}, the function centers each element of
\code{x} by the corresponding element of \code{y}.

Centering is accomplished by the parallel transport specified by XXX (2XXX), by
first computing the inverse square root of \code{y}, and then centering \code{x}
by \code{y.inv.sqrt x y.inv.sqrt}.
}
