% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.dist.R
\name{spd.dist}
\alias{spd.dist}
\title{Distance between two symmetric, positive-definite matrices}
\usage{
spd.dist(x, y, method = "riemannian")
}
\arguments{
\item{x, y}{Symmetric, positive-definite matrices}

\item{method}{The distance measure. See details.}
}
\description{
Function implements several distance measures between symmetric,
positive-definite matrices, not all of which are proper metrics.
}
\details{
Allowable distance measures are
\itemize{
 \item{"frobenius": }{The Frobenius norm of the difference \code{x-y}. Not affinely invariant.}
 \item{"cholesky": }{The Frobenius norm of the difference between the cholesky factors
 of \code{x} and \code{y}. Not affinely invariant.}
 \item{"affine": }{An affinely invariant distance measure, computed by taking
 the Frobenius norm of the projection of \code{y} into the tangent space at \code{x}}
 \item{"riemannian": }{The Riemmanian distance proposed by Barachant, et al. (2013)}
}
}
