% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.estimate.R
\name{spd.estimate}
\alias{spd.estimate}
\title{Regularized covariance estimation}
\usage{
spd.estimate(y, method = "threshold")
}
\arguments{
\item{y}{A data matrix, where rows are observations and columns are variables}

\item{method}{Either soft thresholding ('threshold') or graphical lasso ('sparse')}
}
\value{
A covariance matrix
}
\description{
Function performs regularized covariance estimation using either soft-thresholding
or the graphical lasso, involving l1 penalized estimation of the inverse
covariance (precision) matrix. In either case, the optimal tuning parameter
is selected by cross-validation. Currently, the function's internals are set
to my own default settings, for simplicity.
}
