% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.expmap.R
\name{spd.expmap}
\alias{spd.expmap}
\title{Exponential map onto the space of SPD matrices}
\usage{
spd.expmap(x, p = NULL)
}
\arguments{
\item{x}{A symmetric matrix}

\item{p}{An spd matrix from whose tangent space to perform the projection. If
unspecified, the identity matrix is chosen.}
}
\description{
Function projects one or more vectors onto
the space of SPD matrices space at a point \code{p}
}
\details{
Function uses the exponential map to project a symmetric
matrix onto the space of SPD matrices at a matrix \code{p}.
}
