% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.logmap.R
\name{spd.logmap}
\alias{spd.logmap}
\title{Log-map onto the tangent space}
\usage{
spd.logmap(x, p = NULL)
}
\arguments{
\item{x}{An spd matrix}

\item{p}{An spd matrix on whose tangent space to perform the projection. If
unspecified, the identity matrix is chosen.}
}
\description{
Function projects one or more symmetric, positive-definite matrices onto
the tangent space at a point \code{p}
}
\details{
Function uses the log-map to project a symmetric, positive-definite
matrix onto the tangent space of the positive cone at a matrix \code{p}.
}
