% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.mean.R
\name{spd.mean}
\alias{spd.mean}
\title{Compute the mean of a set of spd matrices}
\usage{
spd.mean(x)
}
\arguments{
\item{x}{A list of symmetric, positive-definite matrices}
}
\value{
The mean of the matrices in \code{x}.
}
\description{
Function computes the mean of a set of symmetric positive-definite matrices
by taking the logarithm of each amtrix, taking the mean, and then transforming
back using the matrix exponential. By default, the function tries to diagonalize
each argument, and then uses the method by Higham (2008) if that fails.
}
\details{
The function uses the \code{logm} and \code{expm} functions in the
\code{expm} package. It first attempts to diagonalize each matrix using
\code{method = 'Eigen'}, switching to \code{meth0d = 'Higham08'} if that fails.
}
