% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.mean.R
\name{spd.mean}
\alias{spd.mean}
\title{Compute the mean of a set of spd matrices}
\usage{
spd.mean(x, method = "euclidean", tol = 0.1, max.iter = 20)
}
\arguments{
\item{x}{A list of symmetric, positive-definite matrices}

\item{method}{The type of mean to compute. See details}

\item{...}{Further arguments. See details}
}
\value{
The mean of the matrices in \code{x}.
}
\description{
Function computes the mean of a set of symmetric positive-definite matrices.
Several methods are implemented, as described in \code{Details}.
}
\details{
Function computes the mean of a set of symmetrix, positive-definite
matrices. Several methods are implemented:
\itemize{
 \item{"euclidean": }{The ordinary arithmetic mean -- the sum of the matrices in x,
 divided by the number of matrices. This is guaranteed to be an spd matrix, but
 does not necessarily preserve the spectral characteristics of the individual matrices.}
 \item{"logeuclidean": }{Computed by taking the arithmetic mean of the logarithms
 of the matrices in \code{x}, and then projecting back onto the space of spd matrices.
 In general, better behaved than the arithmetic mean.}
 \item{"riemannian": }{The geometric mean proposed by Barachant et al. (2013).
 Is approximated iteratively, and accepts two additional argument: \code{max.iter},
 giving the maximum number of iterations (defaults to 20), and \code{tol}, giving
 the maximum Frobenius norm of the approximation error (defaults to 0.1)}
}
}
