% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-transport.R
\name{spd.transport}
\alias{spd.transport}
\title{Parallel transport of a tangent vector}
\usage{
spd.transport(x, to, from, nsteps = 10)
}
\arguments{
\item{x}{A tangent vector (symmetric matrix)}

\item{to}{The SPD matrix to whose tangent space to move x}

\item{from}{The SPD matrix to whose tangent space x belongs}

\item{nsteps}{The number of steps in the geodesic connecting to and from}
}
\value{
A symmetric matrix -- a tangent vector at to.
}
\description{
Parallel transports a vector \code{x} in the tangent space at \code{"from"} to the
tangent space at \code{"to"} using the Schild's ladder algorithm.
}
