% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.whiten.R
\name{spd.whiten}
\alias{spd.whiten}
\title{Whiten a symmetric positive definite matrix}
\usage{
spd.whiten(x, p, dewhiten = F)
}
\arguments{
\item{x, p}{Symmetric, positive-definite matrices}

\item{dewhiten}{If TRUE, reverse the whitening transform}
}
\description{
Approximately whitens an spd matrix x by a matrix p. Specifically,
conjugates x by inverse.sqrt(p), which, assuming that p is relatively close
to x, is close to the identity matrix.
}
